chore: add version implementation for older Go
