+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/madlambda/benchcheck
 
-go 1.18
+go 1.16
 
 require (
 	github.com/google/go-cmp v0.5.7
